--- install.sh
@@ -1,5 +1,0 @@
-git clone https://github.com/Nigma1337/openssh-portable.git
-cd openssh-portable
-autoreconf
-./configure --with-pam
-make install
