--- metadata/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "pam-discord-metadata"
-version = "0.1.0"
-edition = "2021"
-
-[lib]
-name = "pam_discord_metadata"
-crate-type = ["cdylib"]
-
-[dependencies]
-pam-bindings = { version = "0.1.0" }
-hyper = { version = "0.14.20", features = ["full"]}
-tokio = { version = "1.12.1", features = ["full"] }
-futures = { version = "0.3.21"}
